# openapi.yaml
openapi: 3.0.0
info:
  title: Recall Orchestrator API
  description: API for managing product recalls, finding customers, and processing refunds.
  version: 1.0.0
servers:
  - url: https://recall-orchestrator.vercel.app
    description: Vercel Production Server
paths:
...
  /api/issueRefund:
    post:
      summary: Issue a refund (Approve)
      description: Triggers the actual Stripe refund and updates status to 'issued'.
      operationId: issueRefund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refundId:
                  type: string
      responses:
        '200':
          description: Refund issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  stripeRefund:
                    type: object
...